requirements.txt
airbnb==2.0.11
Babel==2.5.3
braintree==3.42.0
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==1.10
django-allauth==0.35.0
django-filter==1.1.0
django-import-export==1.0.0
django-phonenumber-field==2.0.0
django-sendgrid-v5==0.6.87
django-suit==0.2.25
django-twilio==0.9.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.6
jdcal==1.3
Markdown==2.6.11
oauthlib==2.0.6
odfpy==1.3.6
openpyxl==2.5.1
phonenumberslite==8.9.1
psycopg2==2.7.4
PyJWT==1.6.0
python-http-client==3.0.0
python-openid==2.2.5
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
requirements==0.1
sendgrid==5.3.0
six==1.11.0
tablib==0.12.1
twilio==6.10.5
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
# depreciation
# django-endless-pagination
